not great but its getting there

+++ assets/shaders/rtl/rtl_reset.wgsl
@@ -2,7 +2,7 @@
 @group(0) @binding(0)
 var<uniform> uni: f32;
 @group(0) @binding(1)
-var<storage, read_write> lighting_output: array<f32>;
+var<storage, read_write> lighting_output: array<u32>;
 
 
 @compute @workgroup_size(16, 16)
@@ -11,7 +11,7 @@ fn main(@builtin(global_invocation_id) gid : vec3<u32>) {
     let _l = lighting_output[0];
 
     let lightmap_idx = gid.x + (1600 * gid.y);
-    lighting_output[lightmap_idx] = 0.0;
+    lighting_output[lightmap_idx] = 0;
 }
